Allow user to configure number of threads
When calling `build_aomp.sh` there is no control over the number of
threads used, it is calculated in `aomp_common_vars`. For some users
controlling the number of threads might be beneficial. This commit
simply changes the initial name used for discovering the number of
threads and check if `NUM_THREADS` is already defined by the user, if
not follows the same logic as before.