only run sollve 4.5 and 5.0 in run_rocm_test.sh
added EPSDB=1 to invocation of script

QA sees a strange issue with test_target_has_device_addr.c
hangs the gpu, and needs a reboot afterwards

this test is fairly new to sollve suite, its in OpenMP 5.1 language, and
the compiler is giving warnings about unrecognized text at
end, has_device_addr. so its possible the generated code is not
accessing proper addresses and could lead to such a state. these are
part of the 5.1 section of tests which i think the intent of the
run_rocm_test.sh is to not run.