Fix RPATH issue in build_hipamd.sh. (#375)
When using a relocatable install of AOMP, the dynamic
library libamdhip64.so was failing to be resolved
due to not using the proper AOMP RPATH and
was limited to the original install path. Make
use of ROCM_RPATH in build_hipamd.sh to specify
the correct RPATH.

Introduced AOMP_ORIGIN_RPATH_LIST variable in
aomp_common_vars so that this list can be used
in build_hipamd.sh.