[smoke] - Fix clang-262701 test.
The new runtime reports num_threads=1, which causes an issue
with the for loop that sums up values in dist[0]. The upper
bound of the loop is always 2 from the global THREADS. The dist
array is not initilaized and therefore brings in a trash value
for dist[1] and causes a failure in the validation.

We need to ensure the upper bound of the the loop is omp_get_num_threads.