Fixes bug that never checked result of test in check_smoke.sh (#473)
* Fixes bug that never checked result of smoke tests

Introduces new Makefile variable RUNCMD to allow more complex run rules.
When set, this variable is used in `make check` to execute the test.

Some tests were using the 'make run' command to encode a more complex
command for testing. This would never trigger a test failure in
check_smoke.sh as it relies on the 'make check' rule.