[build_hipamd] - Do not export HIP_PATH.
When detecting the hip runtime, llvm looks
in a few different default places for the hip
version file. A new patch now ignores those
default locations and only looks at HIP_PATH
if that variable is set in the environment.
Currently AOMP uses this variable to override
/opt/rocm in files like hipcc and hipvars. We
do not need to export this for the build of
hipamd.