[clone_test] - Add support for fortran-babelstream.
The fortran version is in the same repository,
but a different branch. The fortran-babelstream
directory will use the develop branch instead
of main.

Instead of using the reponame (manifest path)
to clone the repos, we now use repogitname (manifest name).
This permits a duplicate repo to be cloned under a
different directory name to leverage a new branch.