This update adds the new component "flang-legacy" to build_aomp.sh. This creates the binary flang-legacy in the installation bin directory. It is created from flang repo https://github.com/ROCm-Developer-Tools/flang. The flang-legacy component is built before the other three lflang components "pgmath" "flang" and "flang_runtime". There is a single top level source file flang/flang-legacy.cpp that is built using static archives and includes from the frozen 5.5 ROCm lvm compiler. We call this LFL (Last Frozen LLVM).
To find LFL, there is a new prerequiste component called flang-legacy-LFL.
AOMP prerequiste and supplemental components are not part of the compiler
installation. They are only built one time with a call to build_prereq.sh
from build_aomp.sh. The default location is $HOME/local/flang-legacy-LFL
or $AOMP_SUPP/flang-legacy-LFL (if AOMP_SUPP is set). The flang-legacy-LFL
is built by downloading the ROCm 5.5 rocm-llvm package and copying the
archive libraries and includes into $AOMP_SUPP/flang-legacy-LFL. For more
information see the function buildflang-legacy-LFL in build_prereq.sh.

Since flang-legacy is built with static archives, there is NO dependency on
ROCm 5.5 to USE the flang-legacy binary. There is just a BUILD dependency.

The flang-legacy binary is just a command generator for the old legacy
flang which is planned to be deprecated in 2024. The commands use toolchain
arguments created by LFL but they use binaries from the currrent installation.
Mismatched arguments are corrected by logic in flang-legacy.cpp.