Add ability to build openmp using LLVM_ENABLE_RUNTIMES. (#673)
* [build] - Update build scripts for openmp runtime build in llvm.

This change introduces a small patch that is applied
to rocm-device-libs which resideds inside the llvm
mono-repo. Eventually this will be integrated into
amd-stg-open and the small patch can be removed.

This introduces LLVM_ENABLE_RUNTIMES=openmp. In
doing so we have a dependency on rocm-device-libs
and rocr. Rocm-device-libs is built as an external
project along with comgr, which uses the just built
clang. If LIBOMPTARGET_EXTERNAL_PROJECT_HSA_PATH is
passed in for the cmake step then rocr will be
built before openmp, but not installed. build_rocr.sh
will still handle the official build/install steps.

The build_openmp.sh change just disables the normal
openmp library build and is still used for openmp-perf
and openmp-debug.

* Add ability to use RPATH in the llvm openmp build.

* Add correct GFXLIST for llvm openmp build.

* Remove AOMP_ORIGIN_RPATH from build_project.sh.

* Remove unused variable DEXTERNAL_PROJECT_HSA_PATH from build_project.sh

* Add roct build support for build_project.sh.

* Remove external build of comgr in llvm for now.

This was causing some duplicate targets when rebuilding
in the build/llvm-project directory due to comgr
including hip, which in includes AMDDeviceLibs. The
devicelib targets are already present in the build
tree, hence the duplication.