[Smoke] This fixes `make check` (#714)
In `make check` the `RUNCMD` is used within a timeout run. The issue is
that if we have `HSA_XNACK=1` in the beginning of the `RUNCMD` than
timeout will complain that this is not executable.
I moved it to `RUNENV`, which precedes the timeout.
Manually validated that the test still prints out the expected xnack
settings (which it does not test).