[Smoke] Make tests independent of shell environment
This change eliminates the dependency of the two tests on the
environment in which they are executed.
Prior these tests would be marked as fail when running under HSA_XNACK=0
despite the tests producing "expected" results.

The other possibility is to encode different expected results in the
Filecheck strings depending on the HSA_XNACK setting.