Small tweaks to fortran stream_no_teams
I think these tweaks (except perhaps the main size change) are required
to have the correct semantics intended by the functions.

I changed the "dot" function to use a target region, there's no
synch back of the data to array A, B and C from device so the
function will be doing calculation with incorrect data at the
moment (original arrays). So it needs a write back from device
via update or exit  with tofrom (it'd be the latter for now as we
don't actually have update support yet) or we do the calcualtion
on device using target rather than a parallel target as we have no
reduction just now.

Add the allocatable keyword to the input parameters h_A, h_B, h_C of
the "read_arrays" function and explicitly map them as tofrom (could be from
instead to be a little more restrictive):
- The former I'm not sure if it's a bug by the frontend to not indicate that
    you can't pass an allocatable into an argument that doesn't have the
    allocatable keyword on it or if it's legal, if it's legal then it's
    something the allocatable PR currently doesn't handle and I'll likely
    need to consider how to handle it. Which seems like it could be tricky.
  - the latter as an explicit mapping seems neccessary for the read behaviour,
    it could likely be a from mapping as well, instead of tofrom. Currently implicit
    mapping of allocatables doesn't map the data, just the descriptor, not sure
    if the semantics are correct in this case, so I need to look a little more closely
    into it. But in any case if the intent is to read back I think it should likely be an
    explicit from or tofrom map for each allocatable array argument.