[Test] Unifies command executed in run and test (#789)
The issue this patch addresses is that the system constructs the actual
command line for `make run` and `make check` differently.

I have changed this behavior by relying on the `CHECK_COMMAND` in both
rules. This allows us to then also use the exact same variables in both
places. One addition that I need to make was create a new variable that
is used to provide flags for a potentially specified profiler.