Added 4 tests with OpenMP directives in a global constructor.
(1) smoke-limbo/veccopy-ctor-1: OpenMP directive in a global constructor
but not built as a shared library. This currently works.
(2) smoke-limbo/veccopy-ompt-ctor-1: Enabled OMPT for an OpenMP directive
in a global constructor but not built as a shared library. This
currently works.
(3) smoke-fails/veccopy-shared-ctor-1: OpenMP directive in a global
constructor and built as a shared library. This fails an assert, needs
further investigation.
(4) smoke-fails/veccopy-shared-ompt-ctor-1: Enabled OMPT for an OpenMP
directive in a global constructor and built as a shared library. This
fails with the same assert as (3).