[OpenMP] Fix hsa-lazy-queue test by setting ROCR_VISIBLE_DEVICES
Currently, all devices will be initialized during OpenMP plugin creation.
This leads to test fails, when multiple GPUs are present.
(N GPUs -> N+1 calls to hsa_queue_create)

Hence, if ROCR_VISIBLE_DEVICES is not already set, we use '0' as default.
Otherwise, we use the first device that is set as visible.

This will limit test execution to a single GPU and the test will pass as
the expected number of hsa_queue_create calls is observed.