[run_rocm_test.sh] - Add support for detecting skipped tests.
Some smoke tests may be skipped due to not being supported.
For example a USM test on a gfx906. Currently, this test
would be an unexpected fail. If either the run.log or
make-log.txt shows the test was skipped, it will be
removed from the unexpected passing list and a warning
will be issued.

Remove clang-325070 as this bug allowed it to slip past
CI. Will have to look for the fix and re-enable it.