[run_rocm_test] - Now only uses individual passing lists for each ROCm version
Currently, if we are testing ROCm 6.1
the passing lists were combined for previous
rocm releases. This will no longer be the case as
each release will have it's own self contained expected
passing lists. Smoke and smoke-limbo are still expected
to have no failures. Their expected passing lists have
'ALL TESTS SHOULD PASS', which can be added for any
future component as well.

Removed the cleanup at the end of check_smoke.sh
to leave the temp files in place. They are still
cleaned at the beginning of a run.