[build] - Updates for ROCm 6.2 sources (#1020)
* [build] - Updates for ROCm 6.2 sources

* [build] - Add build_rocprofiler-register.sh

* [build] - Update build_rocprofiler-register name in header

* [build] - Additional changes for 6.2 source move

The build of opencl is now dependent on opencl-icd-loader.
This used to be built alongside opencl, but has been deprecated.
The opencl-icd-loader package is built before opencl now, but
has no public sources at the moment. For now we will download
the prebuilt package and store it at /home/estewart/local.