Support separate cmake-only stage for build scripts (#1062)
As part of an experimental project to wrap, isolate and adjust the UI
for these scripts, it's helpful to be able to separate out the running
of the 'cmake' (or 'configure') step from the build and install steps.

This is mostly quite straightforward: we just add a command 'cmake'
(or 'configure') for each script to fit alongside the existing 'nocmake'
and 'install' commands.  One slight difficulty is handling the
flang-legacy script, because it actually builds two components in
succession (llvm-legacy then flang-legacy).  Incremental rebuilds are
more easily handled (in the above-mentioned project) by splitting the
component into two scripts, so that's what this patch does.

OK?