[run_rocm_test] - Reapply package manager logic changes with additional support for EPSDB (#1330)
* [run_rocm_test] - Attempt to download test package first for detection (#1327)

We saw a situation where 'apt' was installed on a CentOS machine
which then tried to use it for the test package download. Instead
of checking package manager version we will attempt to download
the openmp-extras-tests package and if that succeeds then we have
properly identified the package manager.

* [run_rocm_test] - Ensure test package is not downloaded if it already exists

Some cases in CI use a partial mainline build. We do not want to
try and download the test package again. If EPSDB=1 use what is present.