[smoke-fort-fails] Add new identified test causing ICE
The flang compiler appears to mishandle reductions within a target region when
they refer to a single element of an array.

The `target parallel` combined case results in an MLIR verifier error, whereas
the `target` + `parallel` nested case triggers an assertion due to the outlined
parallel region having live-out values.