[build] - Update rccl and rocprofiler-sdk scripts to copy needed files into install
rccl - the .info directory from rocm-core is need in the base install of AOMP
as rccl does a version check based on this file. When we update the rccl
branch in the future there is an option to pass a ROCM_VERSION in via cmake.

rocprofiler-sdk - enhance conditional check for aqlprofile/lib.