[Flang][Test] Add new test-case to use_dev_addr_fort.f90
Adding a new test case I found doesn't work in amd-staging at the moment, but should soon when
I land a new PR. But effectively we crash when we have a map and use_device_addr/ptr to the same
list item on a target at the moment. New patch should allow this.