Various Cherry Picks from 22.0-0 branch to aomp-dev (#1583)
* [build] - Update rocprofiler-sdk cmake prefix path

We need to pass the local directory where prerequisites are
located so that aqlprofile can be found especially when
there is no ROCm installed.

* [build] - copy aqlprofile libs into aomp install directory

rocprofiler-sdk.so depends on aqlprofile libs.

* [build] - Clean up debian packaging script

Remove /usr/bin links to deprecated tools.
Add additional rocmlibs exclusions for aomp debian.

* [rocmlibs][nfc] - fix source_dir typo in cmake echo command

* [build] - Update rccl and rocprofiler-sdk scripts to copy needed files into install

rccl - the .info directory from rocm-core is need in the base install of AOMP
as rccl does a version check based on this file. When we update the rccl
branch in the future there is an option to pass a ROCM_VERSION in via cmake.

rocprofiler-sdk - enhance conditional check for aqlprofile/lib.

* [build] - Add rocprofiler-sdk to tarball exclusions

* [packaging] - fix repo_name/dir_name comparison

There were recent changes that made dir_name a full path. We
now need to get the basename.