Fixed iteam reduction test by replacing num_threads with thread_limit clause. (#1589)
After a recent change, generic SPMD kernels may have their blocksize
reduced by the runtime. So the user code can no longer assume that the
number of threads launched will match the default blocksize. The only
way to ensure that is to add a thread_limit clause to the target
construct.