[smoke-fort] Update kernel-exec-modes test
This test only passed if parallel regions were executed by multiple threads.
This is too restrictive and causes the test to fail if executed in Generic
mode.

Instead of failing, this patch updates the test to print a warning message to
indicate that performance might be impacted when running in single-threaded
mode.