[smoke-fort-dev] Add test for variable size array arguments
This test, which already works when running in SPMD mode (current behavior)
triggers an edge case for Generic mode that causes results to be incorrect.

This is related to the fact that, in that execution mode, a temporary allocated
to store the upper bounds prior to entering the parallel region should use
device shared memory instead of stack memory.